requirements.txt
streamlit
fpdf
python-dotenv
groq
python-docx
PyMuPDF  
docx2txt

